requirements.txt
torch==2.6.0
torchvision==0.21.0
wandb==0.23.1
dvc[s3]>=3.42.0
pandas>=2.2.0
pillow>=10.2.0
scikit-learn>=1.4.0
tqdm>=4.66.1
runpod>=0.6.2
onnx==1.20.0
onnxruntime==1.23.2
